pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wg-federation"
description = "A client/server tool to securely auto-registers peers to a wireguard VPN"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.9"
keywords = ["wireguard", "vpn", "federation", "radius"]
authors = [{ name = "Anuvu DevOps Team", email = "MTL.DevOps.Team@globaleagle.com" }]
maintainers = [{ name = "Anuvu DevOps Team", email = "MTL.DevOps.Team@globaleagle.com" }]
urls = { source = "https://bitbucket.org/GEE_Media/python-wg-federation", container = "https://bitbucket.org/GEE_Media/container-wg-federation" }
# https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Networking",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    'dependency-injector == 4.*',
    'pydantic == 1.*',
    'systemd-python',
    'ipaddr == 2.*',
    'PyYAML == 6.*',
    'deepmerge == 1.*'
]

[project.scripts]
wg-federation = "wg_federation:main"

[project.optional-dependencies]
dev = ['behave', 'pytest>=7', 'pytest-pep8', 'pytest-cov', 'pytest-spec', 'pre-commit', 'pylint', 'tomli', 'mockito']
build = ['setuptools', 'build', 'wheel', 'twine', 'sphinx', 'sphinx_rtd_theme']

[tool.setuptools]
package-dir = { "" = "src" }
license-files = ["LICENSE", "NOTICE"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wg_federation = ["__version__.txt"]

[tool.pytest.ini_options]
spec_test_format = '{result} {docstring_summary}'
minversion = "7.0"
addopts = "-ra"
testpaths = [
    "tests/unit",
]
pythonpath = [
    "src"
]

[tool.mypy]
enable_recursive_aliases = true
